Split language wrappers focal_operation
