+++ source/framework/python/src/algorithm/focal_operation/focal_operation-2.cpp
@@ -1,5 +1,5 @@
-#include "lue/framework/algorithm/value_policies/focal_operation.hpp"
 #include "array_to_kernel.hpp"
+#include "lue/framework/algorithm/value_policies/focal_operation.hpp"
 #include "lue/framework.hpp"
 #include "lue/py/bind.hpp"
 
@@ -9,49 +9,6 @@ using namespace pybind11::literals;
 
 namespace lue::framework {
     namespace {
-
-        class ArithmeticElementBinder
-        {
-
-            public:
-
-                template<Arithmetic Element>
-                static void bind(pybind11::module& module)
-                {
-                    Rank const rank{2};
-
-                    module.def(
-                        "focal_maximum",
-                        [](PartitionedArray<Element, rank> const& array,
-                           pybind11::array_t<BooleanElement> const& kernel) {
-                            return value_policies::focal_maximum(
-                                array, array_to_kernel<BooleanElement, rank>(kernel));
-                        },
-                        "array"_a,
-                        "kernel"_a.noconvert());
-
-                    module.def(
-                        "focal_minimum",
-                        [](PartitionedArray<Element, rank> const& array,
-                           pybind11::array_t<BooleanElement> const& kernel) {
-                            return value_policies::focal_minimum(
-                                array, array_to_kernel<BooleanElement, rank>(kernel));
-                        },
-                        "array"_a,
-                        "kernel"_a.noconvert());
-
-                    module.def(
-                        "focal_sum",
-                        [](PartitionedArray<Element, rank> const& array,
-                           pybind11::array_t<BooleanElement> const& kernel) {
-                            return value_policies::focal_sum(
-                                array, array_to_kernel<BooleanElement, rank>(kernel));
-                        },
-                        "array"_a,
-                        "kernel"_a.noconvert());
-                }
-        };
-
 
         class FloatingPointElementBinder
         {
@@ -135,9 +92,8 @@ namespace lue::framework {
     }  // Anonymous namespace
 
 
-    void bind_focal(pybind11::module& module)
+    void bind_focal2(pybind11::module& module)
     {
-        bind<ArithmeticElementBinder, ArithmeticElements>(module);
         bind<FloatingPointElementBinder, FloatingPointElements>(module);
         bind<IntegralElementBinder, IntegralElements>(module);
     }
